Notification Complete

+++ lib/features/vault/vault_comics_lists.dart
@@ -5,6 +5,7 @@ import 'package:ketemaa/core/Provider/getData.dart';
 import 'package:ketemaa/core/Provider/postData.dart';
 import 'package:ketemaa/core/utilities/app_colors/app_colors.dart';
 import 'package:ketemaa/core/utilities/app_spaces/app_spaces.dart';
+import 'package:ketemaa/core/utilities/common_widgets/customButtons.dart';
 import 'package:ketemaa/core/utilities/shimmer/loading.dart';
 import 'package:ketemaa/features/market/presentation/comic_details.dart';
 import 'package:ketemaa/features/vault/Component/no_data_card.dart';
@@ -85,8 +86,7 @@ class _VaultComicsListState extends State<VaultComicsList> {
                   shrinkWrap: true,
                   itemBuilder: (context, index) {
                     return data.setListModel!.results![index].productDetail!
-                                .type ==
-                            1
+                                .type ==1
                         ? InkWell(
                             onTap: () {
                               /*Get.to(() => ChartExample(id: widget.list![index].id));*/
@@ -434,12 +434,68 @@ class _VaultComicsListState extends State<VaultComicsList> {
                                                 ),
                                                 InkWell(
                                                   onTap: () {
-                                                    postData!.deleteSetList(
-                                                      context,
-                                                      data.setListModel!
-                                                          .results![index].id,
-                                                      requestHeadersWithToken,
-                                                    );
+
+                                                    showDialog(
+                                                        context: context,
+                                                        builder: (context) {
+                                                          return AlertDialog(
+                                                            backgroundColor:
+                                                            AppColors.backgroundColor,
+                                                            shape: const RoundedRectangleBorder(
+                                                              borderRadius: BorderRadius.all(
+                                                                Radius.circular(
+                                                                  20.0,
+                                                                ),
+                                                              ),
+                                                            ),
+                                                            contentPadding:
+                                                            const EdgeInsets.symmetric(
+                                                                horizontal: 20),
+                                                            titlePadding:
+                                                            const EdgeInsets.symmetric(
+                                                                horizontal: 20,
+                                                                vertical: 10),
+                                                            title: Text(""),
+                                                            content: Text(
+                                                              'Do you really want to delete this item?',
+                                                              style: TextStyle(
+                                                                  color: AppColors.textColor,
+                                                                  fontSize: 15),
+                                                            ),
+                                                            actions: <Widget>[
+                                                              CustomButtons(
+                                                                width: Get.width * .2,
+                                                                height: Get.height * .05,
+                                                                onTap: () {
+                                                                  postData!.deleteSetList(
+                                                                    context,
+                                                                    data.setListModel!
+                                                                        .results![index].id,
+                                                                    requestHeadersWithToken,
+                                                                  );
+
+                                                                },
+                                                                text: 'Yes'.toUpperCase(),
+                                                                style: Get.textTheme.button!
+                                                                    .copyWith(
+                                                                    color: AppColors
+                                                                        .textColor),
+
+                                                              ),
+                                                              CustomButtons(
+                                                                width: Get.width * .2,
+                                                                height: Get.height * .05,
+                                                                onTap: () {
+                                                                  Navigator.pop(context);
+                                                                },
+                                                                text: 'Close'.toUpperCase(),
+                                                                style: Get.textTheme.button!
+                                                                    .copyWith(
+                                                                    color: AppColors.textColor),
+                                                              ),
+                                                            ],
+                                                          );
+                                                        });
                                                   },
                                                   child: Icon(
                                                     Icons.delete_forever,
@@ -457,9 +513,11 @@ class _VaultComicsListState extends State<VaultComicsList> {
                               ),
                             ),
                           )
-                        : const NoDataCard(
+                        : Container(
+                          child: const NoDataCard(
                       title: 'Your Wishlist is empty!',
-                    );
+                    ),
+                        );
                   })
               : const LoadingExample(),
         );
